Add reserve server state.
