+++ server/src/main/java/io/atomix/copycat/server/state/InactiveState.java
@@ -82,6 +82,11 @@ class InactiveState extends AbstractState {
   }
 
   @Override
+  protected CompletableFuture<HeartbeatResponse> heartbeat(HeartbeatRequest request) {
+    return Futures.exceptionalFuture(new IllegalStateException("inactive state"));
+  }
+
+  @Override
   protected CompletableFuture<AppendResponse> append(AppendRequest request) {
     return Futures.exceptionalFuture(new IllegalStateException("inactive state"));
   }
