+++ server/src/main/java/io/atomix/copycat/server/state/ReserveState.java
@@ -1,0 +1,238 @@
+/*
+ * Copyright 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License
+ */
+package io.atomix.copycat.server.state;
+
+import io.atomix.catalyst.transport.Connection;
+import io.atomix.copycat.client.error.RaftError;
+import io.atomix.copycat.client.request.*;
+import io.atomix.copycat.client.response.*;
+import io.atomix.copycat.server.CopycatServer;
+import io.atomix.copycat.server.RaftServer;
+import io.atomix.copycat.server.request.*;
+import io.atomix.copycat.server.response.*;
+
+import java.util.concurrent.CompletableFuture;
+
+/**
+ * Reserve state.
+ *
+ * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
+ */
+public class ReserveState extends AbstractState {
+
+  public ReserveState(ServerState context) {
+    super(context);
+  }
+
+  @Override
+  public CopycatServer.State type() {
+    return RaftServer.State.RESERVE;
+  }
+
+  /**
+   * Forwards the given request to the leader if possible.
+   */
+  protected <T extends Request<T>, U extends Response<U>> CompletableFuture<U> forward(T request) {
+    CompletableFuture<U> future = new CompletableFuture<>();
+    context.getConnections().getConnection(context.getLeader()).whenComplete((connection, connectError) -> {
+      if (connectError == null) {
+        connection.<T, U>send(request).whenComplete((response, responseError) -> {
+          if (responseError == null) {
+            future.complete(response);
+          } else {
+            future.completeExceptionally(responseError);
+          }
+        });
+      } else {
+        future.completeExceptionally(connectError);
+      }
+    });
+    return future;
+  }
+
+  @Override
+  protected CompletableFuture<AppendResponse> append(AppendRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(AppendResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<PollResponse> poll(PollRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(PollResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<VoteResponse> vote(VoteRequest request) {
+    context.checkThread();
+    logRequest(request);
+    return CompletableFuture.completedFuture(logResponse(VoteResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<CommandResponse> command(CommandRequest request) {
+    context.checkThread();
+    logRequest(request);
+    if (context.getLeader() == null) {
+      return CompletableFuture.completedFuture(logResponse(CommandResponse.builder()
+        .withStatus(Response.Status.ERROR)
+        .withError(RaftError.Type.NO_LEADER_ERROR)
+        .build()));
+    } else {
+      return this.<CommandRequest, CommandResponse>forward(request).thenApply(this::logResponse);
+    }
+  }
+
+  @Override
+  protected CompletableFuture<QueryResponse> query(QueryRequest request) {
+    context.checkThread();
+    logRequest(request);
+    if (context.getLeader() == null) {
+      return CompletableFuture.completedFuture(logResponse(QueryResponse.builder()
+        .withStatus(Response.Status.ERROR)
+        .withError(RaftError.Type.NO_LEADER_ERROR)
+        .build()));
+    } else {
+      return this.<QueryRequest, QueryResponse>forward(request).thenApply(this::logResponse);
+    }
+  }
+
+  @Override
+  protected CompletableFuture<RegisterResponse> register(RegisterRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(RegisterResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<ConnectResponse> connect(ConnectRequest request, Connection connection) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(ConnectResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<AcceptResponse> accept(AcceptRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(AcceptResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<KeepAliveResponse> keepAlive(KeepAliveRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(KeepAliveResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withLeader(context.getLeader())
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<UnregisterResponse> unregister(UnregisterRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(UnregisterResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<PublishResponse> publish(PublishRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    return CompletableFuture.completedFuture(logResponse(PublishResponse.builder()
+      .withStatus(Response.Status.ERROR)
+      .withError(RaftError.Type.ILLEGAL_MEMBER_STATE_ERROR)
+      .build()));
+  }
+
+  @Override
+  protected CompletableFuture<JoinResponse> join(JoinRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    if (context.getLeader() == null) {
+      return CompletableFuture.completedFuture(logResponse(JoinResponse.builder()
+        .withStatus(Response.Status.ERROR)
+        .withError(RaftError.Type.NO_LEADER_ERROR)
+        .build()));
+    } else {
+      return this.<JoinRequest, JoinResponse>forward(request).thenApply(this::logResponse);
+    }
+  }
+
+  @Override
+  protected CompletableFuture<LeaveResponse> leave(LeaveRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    if (context.getLeader() == null) {
+      return CompletableFuture.completedFuture(logResponse(LeaveResponse.builder()
+        .withStatus(Response.Status.ERROR)
+        .withError(RaftError.Type.NO_LEADER_ERROR)
+        .build()));
+    } else {
+      return this.<LeaveRequest, LeaveResponse>forward(request).thenApply(this::logResponse);
+    }
+  }
+
+  @Override
+  protected CompletableFuture<HeartbeatResponse> heartbeat(HeartbeatRequest request) {
+    context.checkThread();
+    logRequest(request);
+
+    if (context.getLeader() == null) {
+      return CompletableFuture.completedFuture(logResponse(HeartbeatResponse.builder()
+        .withStatus(Response.Status.ERROR)
+        .withError(RaftError.Type.NO_LEADER_ERROR)
+        .build()));
+    } else {
+      return this.<HeartbeatRequest, HeartbeatResponse>forward(request).thenApply(this::logResponse);
+    }
+  }
+
+}
